+++ cli/csstools-cli/CHANGELOG.md
@@ -3,12 +3,16 @@
 ### Unreleased (minor)
 
 - Emit warnings from plugins
-- Added `@csstools/postcss-is-pseudo-class`
+- Added `@csstools/postcss-font-format-keywords`
 - Added `@csstools/postcss-hwb-function`
+- Added `@csstools/postcss-is-pseudo-class`
+- Added `@csstools/postcss-normalize-display-values`
 
 ```sh
-npx -y @csstools/cli postcss-is-pseudo-class
+npx -y @csstools/cli postcss-font-format-keywords
 npx -y @csstools/cli postcss-hwb-function
+npx -y @csstools/cli postcss-is-pseudo-class
+npx -y @csstools/cli postcss-normalize-display-values
 ```
 
 ### 1.0.0 (January 2, 2022)